Optimise the performance of lesson plan edit page
